Get exportNlogo working with optional turtleShapes, linkShapes, info
